reorganize test Data Layers integration to match test mod
