+++ Pathoschild.Stardew.sln
@@ -54,7 +54,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TractorMod", "TractorMod\Tr
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CentralStation", "CentralStation\CentralStation.csproj", "{DE72EBFD-2E30-47C1-A941-82E8922DED6B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleDataLayer", "ExampleDataLayer\ExampleDataLayer.csproj", "{D1400E42-46B2-40B7-A5AC-43C2B9A39687}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestDataLayersMod", "TestDataLayersMod\TestDataLayersMod.csproj", "{D1400E42-46B2-40B7-A5AC-43C2B9A39687}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
@@ -165,6 +165,7 @@ Global
 		{EC5F07FF-A84F-4088-9FA2-DAE66046A093} = {5950ED4C-2284-4253-9583-BB7811FAF5EE}
 		{EF4F1A38-6601-4441-B7B1-E9C26B312192} = {5950ED4C-2284-4253-9583-BB7811FAF5EE}
 		{729DB4C6-AD85-4BAF-B54A-DF6D534E50FF} = {5950ED4C-2284-4253-9583-BB7811FAF5EE}
+		{D1400E42-46B2-40B7-A5AC-43C2B9A39687} = {5950ED4C-2284-4253-9583-BB7811FAF5EE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {EF4A4878-3839-48F1-B4C1-43D0570BCCB4}
