+++ TestDataLayersMod/TestDataLayersMod.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <Version>1.17.1</Version>
-    <RootNamespace>Pathoschild.Stardew.ExampleDataLayer</RootNamespace>
+    <Version>1.0.0</Version>
+    <RootNamespace>Pathoschild.Stardew.TestDataLayersMod</RootNamespace>
 
     <TranslationClassBuilder_AddGetByKey>true</TranslationClassBuilder_AddGetByKey>
     <TranslationClassBuilder_AddKeyMap>true</TranslationClassBuilder_AddKeyMap>
