+++ TestDataLayersMod/manifest.json
@@ -1,0 +1,9 @@
+﻿{
+    "Name": "Test Data Layers Integration",
+    "Author": "Pathoschild",
+    "Version": "1.0.0",
+    "Description": "A dev mod used to test the Data Layers integration.",
+    "UniqueID": "Pathoschild.TestDataLayersMod",
+    "EntryDll": "TestDataLayersMod.dll",
+    "UpdateKeys": []
+}
